# Replace dask-cudf set_index code with upstream