@mrocklin 2019-08-19
I understand the need of a wrapper in order to handle the disk/tasks ambiguity.  However I'm not sure why the rest of the lines of code there around sorted, inplace, and so on, were necessary.  Why was this?