@rjzamora 2019-08-20
> However I'm not sure why the rest of the lines of code there around sorted, inplace, and so on, were necessary. Why was this?

Sorry about that -  I simplified this.