@mrocklin 2019-08-21
Fixes https://github.com/rapidsai/cudf/issues/2292