@mrocklin 2019-08-21
I'm happy to see how simple this is.  

I imagine that @kkraus14 has a question on if this slows things down or not.  if it's easy @rjzamora can you provide a couple of examples on performance difference?  Perhaps one with only two columns and one with ten?

```python
df = dask.datasets.timeseries(...)

%%time
df.set_index(...)
df.persist()
wait(df)  # only if you're using a dask.distributed Client
```