@rjzamora 2019-08-21
> I imagine that @kkraus14 has a question on if this slows things down or not. if it's easy @rjzamora can you provide a couple of examples on performance difference? Perhaps one with only two columns and one with ten?

We are see minor/moderate speedups with the upstream `set_index` ("new version"). Using `dask.datasets.timeseries` with 2,592,000 rows (30 partitions)...

**Non-Distributed, 2 Columns** (4.4x Speedup): 

Old Version:  ~48.9 s
New Version:  ~11 s

**Non-Distributed, 10 Columns** (4x Speedup): 

Old Version:  ~81 s
New Version:  ~20.3 s

**DISTRIBUTED, 2 Columns** (5x Speedup): 

Old Version:  ~36.4 s
New Version:  ~7.3 s

**DISTRIBUTED, 10 Columns** (5.4x Speedup): 

Old Version:  ~52.3 s
New Version:  ~9.7 s